fetch: Download all sources defined in the spec file
The make pattern rule which calls fetch only fires once, even if the
rule defines multiple targets.   Make does not provide a variable to
access all the targets, only the one which happened to be the first to
need to be rebuilt.   To handle a spec file which needs several sources
we must download all of them in one invocation of fetch.

Signed-off-by: Euan Harris <euan.harris@citrix.com>