pin: Combine pinned source versions to create new version
The new version now will concatenate the source number being overriden with the
version of that source target.

For example, if package "my-package" had two sources and both were
independently pinned, the resulting version would look like:

  s0+<descr0>_s1+<descr1>

Where descrN is the result of a git-describe on the source pin. An example
might look like:

  my-package-s0+0.1+1+g3a73c41_s1+0.1-1.el6.x86_64.rpm

We can now tell exactly what source was used. In the above example, we can
infer that Source0 for this package came from a commit with hash 3a73c41 that
was exactly one commit ahead of the tag '0.1' and Source1 came from a a commit
that was tagged with '0.1'.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>