pin: Use release tag in spec file for snapshot version
This is inline with the fedora packaging guidlines here:

https://fedoraproject.org/wiki/Packaging:NamingGuidelines?rd=Packaging/NamingGuidelines#SnapshotPackages

This patch makes sure the tarball unpacks with the old version by using the
original version from the spec file as the --prefix for the git archive. But it
does rename the final resulting tarball so we can distinguish it in the
SOURCES directory.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>