Makefile: Improve locking around repo metadata
There can be a race between mock consuming repository
metadata and createrepo producing it.   This is especially
apparent when planex-cache is used instead of mock - it also
consumes the repository metadata, but cycles so quickly that
the likelihood of a race is much greater than with conventional
mock.

This changeset wraps mock/planex-cache with a shared (a.k.a. read)
lock on the RPMS directory.   The createrepo step already takes an
exclusive (a.k.a. write) lock on that directory - this changeset
makes that more explicit by passing --exclusive, which is the default
for flock.

This change should prevent races between mock and createrepo, but
it may come at the cost of some performance.   An atomic
compare-and-swap wrapper around createrepo might be better, but
the interaction with yum would still be tricky.

Signed-off-by: Euan Harris <euan.harris@citrix.com>