fetch: Restore ability to fetch individual sources
This reverts commit 15cdc0c2b96d2e77185cdeb645b8eec925f326d8.

The fetch rules we generate are not pattern rules, so fetch
is called once for each source tarball.   This means that the
reasoning behind commit 15cdc0 is flawed.   Furthermore,
the change to download to a temp file causes concurrent
fetch runs to race - the first run will move the temp file
aside, then the second will try and fail.

Signed-off-by: Euan Harris <euan.harris@citrix.com>