planex-cache: Don't fail if the binary package is already in the cache
planex-cache can fail if another racing build writes a binary RPM to the
cache between the time when cache checks the cache and tries to write
back to it after building the package.   If the package is already in the
cache, we can just throw away our local build.

Fixes #238.

Signed-off-by: Euan Harris <euan.harris@citrix.com>