cache: Make sure we call /usr/bin/mock instead of /usr/sbin/mock
Mock should be run by a normal user account calling /usr/bin/mock.
If the user's path causes /usr/sbin/mock to be picked up instead,
it will exit with a warning but with a 0 exit code.
Subprocess.Popen will not detect this as an error and the build will
appear to succeed without building anything.

Signed-off-by: Euan Harris <euan.harris@citrix.com>