makesrpm: Strip leading path components from spec file
If makesrpm is called with a pinned spec file, the spec file path will
start with PINS, not SPECS.   This causes makesrpm to fail, as it tries
to copy the spec file to $tmpdir/PINS/foo.spec, which does not exist.
This change strips the path from the spec file when generating the destination
path.

Signed-off-by: Euan Harris <euan.harris@citrix.com>