extract: prepend manifest of sources (and branch variable)
Add a comment block to the start of the spec file which lists the full
URLs of external sources.

Patchqueues contain spec files with source URLs. These currently point
to a specific branch. During spec file rewrite define a variable
'branch' which is derived from the link file which can be used in
place of a hard-coded branch name.

Signed-off-by: Simon Rowe <simon.rowe@eu.citrix.com>