Docker: change "build" user creation to avoid scanning large directories
Running the "usermod" command to change the uid of a user has the side
effect of changing the uid in all files in the home of the user to match
the new uid.

When ran in a directory having a large number of files on a slow fs (NFS)
it can take awfully long time.

The planex docker image uses usermod to set up the "build" user inside the
container with a right uid, so it has the issue described above.

With this change the "build" user is created with the right uid in the
first place, so "usermod" is not needed, thus fixing the performance issue.


Signed-off-by: Istvan Szekeres <istvan.szekeres@citrix.com>