mock: Automatically add loopback repository stanza
In a temporary directory, create a new mock config
based on the config passed in on the command line,
with a suitable loopback Yum repository stanza.
Make sure that the last-modified time of the config
file is the same as that of the input config - if
the config file is newer than the mock chroot cache
then the cache will be rebuilt.   The packages in the
loopback repository are not copied into the mock
cache so it is safe to trick mock into not rebuilding
it.  If the original mock config is changed, the cache
will be rebuilt correctly.

Also copy logging.ini and site-defaults.cfg, ensuring
that their last-modified times are preserved.

Signed-off-by: Euan Harris <euan.harris@citrix.com>