makesrpm: Do not extract all files in source directories
Some packages keep sources and patches in the top level directory of
the archive, along with the spec file.    Extracting all files in this
directory will overwrite the processed spec file, losing the expanded
patch queue.   The SRPM will still build, but none of the patchqueue
patches will be present in it.

Signed-off-by: Euan Harris <euan.harris@citrix.com>