planex-clone: Clone patchqueue-based packages
Without --jenkins argument:
 * Clone base repository and check out appropriate branch
 * Clone patchqueue repository
 * Symlink patchqueue repository into .git/patches in base repository
 * Create necessary guilt state
 * guilt push all patches

This sets up an environment in which a developer can work on the
patchqueue and build an RPM from it.

With --jenkins argument:
 * Print Jenkinsfile fragment to clone patchqueue repository

This only clones the patch queue.   The underlying repository
tarball will be fetched as normal, the patch queue will be
archived and added to the SRPM.

Signed-off-by: Euan Harris <euan.harris@citrix.com>