planex-build-mock: Add --init option to warm the mock cache
When mock is first run, it builds a chroot image by downloading
and installing the RPMs specified in the 'chroot_setup_command' in
mock/default.cfg and saving the image as a tarball.   If several mock runs
are started at the same time and no root cache exists, they will all start
building one and will overwrite each others' tarballs as they complete.
This can lead to the root tarball being regenerated several times in the
early phase of a concurrent build.  To avoid this, we can call 'mock
init' before starting the build.  This populates the root cache once,
and the actual build jobs all use this cache rather than rebuilding it.

This change adds a flag to planex-build-mock which makes this possible.

Signed-off-by: Euan Harris <euan.harris@citrix.com>