patchqueue: Fix patch generation with complex history
git-format-patch is intended for use with simple linear histories which
means that it does not generate a consistent set of patches when the
history involves merges.  As an alternative, make a list of patches
similar to how git-format-patch does but include the diffs for merges
and ensure that the patch series is linear by following the first parent
only.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>