clone: Handle pins to branches, tags and commit hashes correctly
If a repository is pinned to a tag or commit hash, we must create
a local branch at that commit and check it out.  In this context,
`git checkout -B <branch|tag>` will create a new branch with the
appropriate name, but the branch will point at head on master.

Signed-off-by: Euan Harris <euan.harris@citrix.com>