Docker: Pass SSH_AUTH_SOCK through to sudo session in container
Cloning repositories over ssh relies on having access to credentials.
Mount the SSH_AUTH_SOCK into the container as a volume and pass the
SSH_AUTH_SOCK environment variable through.   To do this, we need
to use sudo rather than su to become the build user, as sudo allows
us to set individual environment variables whereas for su environment
preservation is all-or-nothing.

Signed-off-by: Euan Harris <euan.harris@citrix.com>