Docker: Translate entry script into Python
The entrypoint script in Dockerfile is the first process to run inside the
container.   It creates the build user and group, setting the UID and GID
to match the owner of the directory mounted at /build, so that the files
are readable within the container, becomes the bulid user and then either
runs an interactive shell or a command passed on the commandline.

As the script becomes more complicated, keeping track of the different
arguments passed between layers of shells is becoming more difficult
and fragile.   Rewriting this script in Python gives us more powerful
tools to keep track of what is going on.

Signed-off-by: Euan Harris <euan.harris@citrix.com>