CP-21573: Exclude '*-release' packages from the 'mock-loopback' repo
All '<oem>-release' packages obsolete the 'centos-release' package,
and provide it as a capability. This package, however, is a runtime
dependency of 'systemd'.

If more that 1 '<oem>-release' package exists in the 'mock-loopback'
repo and we try to build a package that requires 'centos-release',
yum will try to install them all, effectively breaking the build,
as all '<oem>-release' packages conflict with one another.

Therefore, we exclude all such packages from 'mock-loopback' repo,
leaving yum to use 'centos-release' for all build time dependencies.

Signed-off-by: Kostas Ladopoulos <konstantinos.ladopoulos@citrix.com>