makesrpm: Do not rewrite %autosetup rules in spec files
GitWeb synthetic tarballs do not unpack into a top level directory with
the name RPM expects.   To work around this, we rewrite %autosetup rules
in spec files.   GitHub generates tarballs which are acceptable to RPM
by default and BitBucket allows the top level directory name to be
specified in the tarball URL, so this rewrite is no longer required.

Signed-off-by: Euan Harris <euan.harris@citrix.com>