mock: ensure that build output is always visible
mock behaves differently whether stderr is a tty or not:
https://github.com/rpm-software-management/mock/blob/c59581ac78098b2b58b3a30a4a5d1f7338017e6b/py/mockbuild/util.py#L944
'make output-sync=target' redirects the output, so that stderr is no longer a tty, causing the loss of build output.

Restore mock loglevel to normal, and use pty.spawn to turn mock's stderr into a pseudo-tty.
For compatibility with python2.7 explicitly check the exitcode of the child process, because we want to propagate build errors to the parent process.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>