fetch: Match requested sources correctly
Previously, we searched for a source path in the spec file which ended
with the basename of the requested file.  This failed if the package
declared a different source with a longer name but the same suffix e.g.
kdump and xs-kdump.

We now compare the basename of the requested source path to the basename
of the source declared in the spec file.

Signed-off-by: Euan Harris <euan.harris@citrix.com>