depend: Do not make an SRPM depend on a link if a pin is also present
A pin should override a link.   There is no need for an SRPM to depend
on a link file if it also depends on a pin.   Currently this does not
cause problems because planex-make-srpm only uses the first pin or link
it finds on the command line, and planex-depend prints the pin dependency
first.

Signed-off-by: Euan Harris <euan.harris@citrix.com>