Makefile.rules: Specific rule to fetch patchqueues is not needed
Previously, patchqueue had to be downloaded before dependencies were
generated, so that spec files could be extracted from them.   This was
done by a pattern rule which downloaded a patchqueue for each link file.
As all specs are now stored in the specs repository, dependencies can be
generated without downloading the patchqueues.   This allows planex-depend
to produce dependency rules for the patchqueues which will trigger the
generic fetch rule, removing the need for a specific pattern rule.

Signed-off-by: Euan Harris <euan.harris@citrix.com>