depend: Depend on local sources if link/pin does not provide them
If a linked (or pinned )spec file refers to local sources and the link
file provides them, we do not need to print a dependency between the
SRPM and the source because this will be covered by the dependency on
the tarball built from the link.  If the link does not provide sources,
assume that they can be found in the specfile repository.

Signed-off-by: Euan Harris <euan.harris@citrix.com>