mock: Fix RPMS top-level link
After 4fc2f628, links are created to the x86_64 and noarch subdirectories
instead of the RPMS directory.   This happens because we now call `ln
-sf $(@D)`.  `$(@D)` is the directory part of the target and because
we split the contents of RPMS by architecture the path points to
$(PWD)/_build/RPMS/{x86_64, noarch}.    We can't just take `dirname $(@D)`
because another setup might not split the contents of RPMS;  we could
ask RPM to tell us _rpmdir but since we would like to get rid of these
symlinks entirely I think returning to something more like the previous
hard-coded behaviour is the best option.

Signed-off-by: Euan Harris <euan.harris@citrix.com>