depend: Make a package depend on its Requires as well as its BuildRequires
To build package A, mock needs to install all the package's build-time
requirements (BuildRequires). It does not need to install the package's
runtime requirements (Requires). However, to build package B, which
depends on A, mock must install A's runtime requirements. This causes
problems if A's runtime requirements come from the same spec repository as
A itself. planex-depend does not produce a Makefile dependency which will
cause A's runtime dependencies to be build before B is built, so it is
the luck of the draw whether those packages exist at the time B is built.

Technically, planex-depend should make B have a build dependency on
A's runtime dependencies, however it is easier to make A depend on its
own runtime dependencies, even though they are not strictly required in
order to build it.

As with BuildRequires, if a runtime dependency is not provided by any
of the specs we are building, we will assume that the dependency comes
from the environment and no Makefile rule will be created.

Signed-off-by: Euan Harris <euan.harris@citrix.com>