makesrpm: Avoid unlink races in ln -sf
A race can occur when calling ln -sf in older versions of GNU coreutils (and
in busybox):

   http://git.savannah.gnu.org/gitweb/?p=coreutils.git;a=commit;h=376967889ed7ed561e46ff6d88a66779db62737a

We seem to be hitting this race because we use ln -sf to create the
RPMS and MANIFESTS top-level symlinks for every target.   Two racing
ln invocations try to remove the existing symlink, one succeeds and the
other fails with the following message:

  ln: cannot remove './RPMS': No such file or directory

This does not happen all the time because the first ln may have created
its link, which the second then deletes.

This commit switches back to the previous behaviour of creating these
symlinks at the start of the build.   This has caused problems in the
past (bug 211) so we intend eventually to remove these links altogether.

Signed-off-by: Euan Harris <euan.harris@citrix.com>