spec: Use RPM object's sprintf to expand tags in _build_name_fmt
The capitalized placeholders %{ARCH}, %{NAME}, %{VERSION} and %{RELEASE}
in the %_build_name_fmt macro refer to RPM tag names, not names of other
macros.   The rpm.tagnames dictionary gives the names of all known tags
and the integers they map to internally.  A package's RPM object exposes
the package's tags as a dictionary, so hdr['name'] or hdr['NAME'] gets
the value of the NAME tag.

The correct way to expand tags is to use the RPM object's sprintf()
method.   We should also set the corresponding RPM macros such as %name
and %version to the appropriate values so that actual macros are expanded
consistently.

See https://bugzilla.redhat.com/show_bug.cgi?id=999778 for more information.

Signed-off-by: Euan Harris <euan.harris@citrix.com>