planex-pin: implement the new planex-pin
The CLI is now completely rewritten. Two mutually exclusive flags allow
to write on a specific path ('-o') or on the automatically generated
"PINS/$package_name.pin" path ('-w'). The default behaviour is still to
print on stdout only.

By default planex-pin $pkgname will load $pkgname.spec and $pkgname.lnk
and generate a pin file with the current state of affairs from these two.
Two mutually exclusive flags can be used to create a fresh pin with a
single Source0 as source of truth, or a pin file where the PatchQueue us
replaced by a custom PatchQueue.  These are documented in the command
help.

Signed-off-by: Marcello Seri <marcello.seri@citrix.com>