clone.py: extend clone behaviour
Previously we were creating handmade planex/ branches if
by checking out the commitish we would have ended in
detached head state. Now this will happen only if we
use the --apply-patchqueue or --apply-repatched flags.
In that case we need to be in a proper branch to be
able to do the various patch applications.

Signed-off-by: Marcello Seri <marcello.seri@citrix.com>