CA-247468 Make makesrpm.py more self-contained
In order to avoid issues where "rpmbuild -bs" as invoked by makesrpm.py
is trying to mess with the RPM _topdir directories as used by other
parts of the process at the same time, separate that invocation of
"rpmbuild -bs" into its own separate _topdir (it was already partway
there).

While doing this, separate all the "--define" arguments into individual
makefile variables and supply them only where they are really needed. We
are supplying "--define _topdir" to the mock wrapper because it does
extra repository creation work which relies on expanding "%_rpmdir", but
then we have explicit code to avoid passing it on to mock because it
would break things. We *were* passing "%_sourcedir" on to mock and
getting away with it because it was ignoring us on a rebuild operation.

This reduces the risk that new arguments added to a "used everywhere"
variable will break things.